pyproject.toml
[project]
name = "qpdk"
version = "0.3.0"
description = "Quantum-RF generic PDK"
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { file = "LICENSE" }
authors = [{ name = "gdsfactory", email = "contact@gdsfactory.com" }]
keywords = ["python"]
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = ["doroutes>=0.2.0", "gdsfactory~=9.39.0"]

[project.optional-dependencies]
models = [
  "gplugins[meshwell]>=2.0.1",
  "jaxellip",
  "optuna",
  "pymablock @ git+https://gitlab.kwant-project.org/qt/pymablock.git@c441ef5d3d95b7823a08cf81da686876ae2e5231",
  "ray[tune]; sys_platform != 'win32'",
  "sax>=0.16.11",
  "scikit-rf",
  "scqubits",
  "sympy",
  "polars",
]

[dependency-groups]
dev = [{ include-group = "lint" }, { include-group = "test" }]
docs = [
  "jinja2",
  "jupyter-book>=1.0.4,<2",
  "jupytext",
  "matplotlib",
  "pydata-sphinx-theme",
  "sphinxcontrib-katex",
  "sphinxcontrib-svgbob",
]
github = ["genbadge[coverage]", "setuptools"]
lint = ["prek", { include-group = "stubs" }]
stubs = ["scipy-stubs"]
test = [
  "hypothesis",
  "jsondiff",
  "pytest>=9.0.0",
  "pytest-cov",
  "pytest-github-actions-annotate-failures",
  "pytest_regressions",
  "pytest-xdist",
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[tool.codespell]
ignore-words-list = "te, te/tm, te, ba, fpr, fpr_spacing, ro, nd, donot, scheme, Commun, Ue, inout"
skip = "bibliography.bib"

[tool.gdsfactoryplus]
name = "qpdk"

[tool.gdsfactoryplus.drc]
timeout = 300

[tool.gdsfactoryplus.sim.x]
max = 20  # 20GHz
min = 1  # 1GHz
name = "f"
num = 3000

[tool.gdsfactoryplus.pdk]
name = "qpdk"

[tool.interrogate]
docstring-style = "google"
exclude = [
  "docs/_build/**/*",
  "qpdk/**/__init__.py",
  "notebooks/**",
  "docs/notebooks/*",
  "qpdk/samples/*.py",
  "tests/**",
]
ext = []
fail-under = 100
ignore-nested-classes = true
ignore-nested-functions = true
ignore-regex = ["^test_.*"]
omit-covered-files = true
verbose = 2

[tool.pylsp-mypy]
enabled = true
live_mode = true
strict = true

[tool.pyrefly]
project-excludes = [
  "**/__pycache__",
  "**/.*",
  "docs/**",
  "build/**",
  "dist/**",
  "*.egg-info/**",
  ".venv/**",
  # Exclude samples that may have experimental code
  "qpdk/samples/*.py",
  "qpdk/samples/**/*.py",
  "notebooks/src/*.py",
  "notebooks/src/**/*.py",
]
project-includes = ["**/*.py"]

[tool.pyrefly.errors]
bad-argument-type = "ignore"
# Configure specific error types - start permissive and tighten over time
missing-argument = "ignore"
missing-module-attribute = "ignore"
# Not all gdsfactory overloads are seem to be properly typed
no-matching-overload = "ignore"
not-a-type = "warn"
read-only = "ignore"
# The cast() calls are needed for pyrefly <0.55 compatibility (pre-commit hook uses 0.46.3).
# TODO: remove this when the pre-commit hook is updated to pyrefly >=0.55.0.
redundant-cast = "ignore"

[tool.coverage.run]
source = ["qpdk"]

[tool.coverage.report]
exclude_lines = [
  "if __name__ == .__main__.",
  "pragma: no cover",
]

[tool.pytest]
addopts = ["--tb=short"]
markers = ["skip_windows: skip test on Windows"]
norecursedirs = [
  "*.egg",
  ".*",
  "CVS",
  "_darcs",
  "build",
  "dist",
  "extra/*.py",
  "node_modules",
  "venv",
  "{arch}",
]
python_files = ["notebooks/*.ipynb", "qpdk/*.py", "tests/*.py"]
testpaths = ["qpdk/", "tests"]

[tool.ruff]
fix = true

[tool.ruff.format]
exclude = ["*.ipynb"]

[tool.ruff.lint]
exclude = ["*.ipynb"]
ignore = [
  "E501",  # line too long, handled by black
  "B008",  # do not perform function calls in argument defaults
  "C901",  # too complex
  "B905",  # `zip()` without an explicit `strict=` parameter
  "C408",  # C408 Unnecessary `dict` call (rewrite as a literal)
  "RUF001",  # ambiguous unicode characters (incl. lots of math)
  "RUF002",  # ambiguous unicode characters (incl. docstring greek letters)
  "RUF003",  # -||- but for comments
]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "C",  # flake8-comprehensions
  "B",  # flake8-bugbear
  "D",  # pydocstyle
  "UP",  # pyupgrade
  "T10",  # flake8-debugger
  "S",  # flake8-bandit
  "PT",  # flake8-pytest-style
  "SIM",  # flake8-simplify
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "Q",  # flake8-quotes
  "SLOT",  # flake8-slots
  "ARG",  # flake8-unused-arguments
  "PTH",  # flake8-use-pathlib
  "PD",  # pandas-vet
  "FLY",  # flynt
  "NPY",  # numpy
  "PERF",  # perflint
  "RUF",  # ruff
]

[tool.ruff.lint.per-file-ignores]
".github/write_cells.py" = [
  "S701",
]  # Jinja2 autoescape not needed for RST templates
"notebooks/**/*.py" = ["D100"]
"qpdk/**/__init__.py" = ["F403"]  # allowing star imports to aggregate cells
"qpdk/samples/**/*.py" = [
  "D100",
]  # no public module docstring for sample scripts
"tests/**" = ["S101", "D102", "D103"]
"qpdk/models/__init__.py" = ["E402"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.tombi.schema]
strict = false

[tool.uv]
default-groups = ["dev", "docs", "lint", "stubs", "test"]

[tool.uv.build-backend]
module-root = ""
source-exclude = ["tests/**", "docs/**"]
